Remove utils.structure module
