+++ openforcefield/utils/utils.py
@@ -881,3 +881,75 @@ def recursive_attach_unit_strings(smirnoff_data, units_to_attach):
         pass
 
     return smirnoff_data
+
+def get_molecule_parameterIDs(molecules, forcefield):
+    """Process a list of molecules with a specified SMIRNOFF ffxml file and determine which parameters are used by
+    which molecules, returning collated results.
+
+    Parameters
+    ----------
+    molecules : list of openforcefield.topology.Molecule
+        List of molecules (with explicit hydrogens) to parse
+    forcefield : openforcefield.typing.engines.smirnoff.ForceField
+        The ForceField to apply
+
+    Returns
+    -------
+    parameters_by_molecule : dict
+        Parameter IDs used in each molecule, keyed by isomeric SMILES
+        generated from provided OEMols. Each entry in the dict is a list
+        which does not necessarily have unique entries; i.e. parameter IDs
+        which are used more than once will occur multiple times.
+
+    parameters_by_ID : dict
+        Molecules in which each parameter ID occur, keyed by parameter ID.
+        Each entry in the dict is a set of isomeric SMILES for molecules
+        in which that parameter occurs. No frequency information is stored.
+
+    """
+    warnings.warn(DEPRECATION_WARNING_TEXT, PendingDeprecationWarning)
+
+    from openforcefield.topology import Topology
+    # Create storage
+    parameters_by_molecule = dict()
+    parameters_by_ID = dict()
+
+    # Generate isomeric SMILES for each molecule, ensuring all molecules are unique
+    isosmiles = [ molecule.to_smiles() for molecule in molecules ]
+    already_seen = set()
+    duplicates = set(smiles for smiles in isosmiles if smiles in already_seen or already_seen.add(smiles))
+    if len(duplicates) > 0:
+        raise ValueError("Error: get_molecule_parameterIDs has been provided a list of oemols which contains some duplicates: {}".format(duplicates))
+
+    # Assemble molecules into a Topology
+    topology = Topology()
+    for molecule in molecules:
+        topology.add_molecule(molecule)
+
+    # Label molecules
+    labels = forcefield.label_molecules(topology)
+
+    # Organize labels into output dictionary by looping over all molecules/smiles
+    for idx in range(len(isosmiles)):
+        # Pull smiles, initialize storage
+        smi = isosmiles[idx]
+        parameters_by_molecule[smi] = []
+
+        # Organize data for this molecule
+        data = labels[idx]
+        for force_type in data.keys():
+            for atom_indices, parameter_type in data[force_type].items():
+
+                pid = parameter_type.id
+                # Store pid to molecule
+                parameters_by_molecule[smi].append(pid)
+
+                # Store which molecule this pid occurred in
+                if pid not in parameters_by_ID:
+                    parameters_by_ID[pid] = set()
+                    parameters_by_ID[pid].add(smi)
+                else:
+                    parameters_by_ID[pid].add(smi)
+
+    return parameters_by_molecule, parameters_by_ID
+
